Created types folder and moved props
