+++ src/index.ts
@@ -1,22 +1,11 @@
-export {
-  useAccounts,
-  useBanks,
-  useMerchantTags,
-  useMerchants,
-  usePaymentRequest,
-  usePaymentRequestMetrics,
-  usePaymentRequests,
-  useTransactions,
-  useUser,
-  useUserPaymentDefaults,
-} from './hooks'
+export * from './hooks'
 
 export * from './clients'
 
-export * from './responseTypes/ApiResponses'
+export * from './types/ApiResponses'
 
-export * from './responseTypes/ApiRequests'
+export * from './types/ApiRequests'
 
-export * from './responseTypes/Enums'
+export * from './types/Enums'
 
-export * from './props/props'
+export * from './types/props'
